pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "cloud-agents"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "ollama>=0.4.0",
    "chromadb>=0.5.20",
    "sentence-transformers>=3.1.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-ollama>=0.2.0",
    "databricks-sdk>=0.25.0",
    "snowflake-connector-python>=3.8.0",
    "boto3>=1.35.0",
    "gradio>=5.0.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "tenacity>=8.3.0",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
]

[tool.setuptools.packages.find]
include = ["agents*", "tools*", "rag*", "orchestrator*", "eval*", "ui*"]
exclude = ["tests*", "infra*"]
